Many fixes
