+++ packages/client/lib/sync/sync.ts
@@ -248,22 +248,6 @@ export abstract class Synchronizer {
     }
   }
 
-  /**
-   *
-   * Processes `NEW_BLOCK` announcement from a peer and inserts into local chain if child of chain tip
-   * @param blockData `NEW_BLOCK` received from peer
-   */
-  async handleNewBlock(block: Block) {
-    const chainTip = (await this.chain.getLatestHeader()).hash()
-    // If block parent is current chain tip, insert block into chain
-    if (chainTip.toString('hex') === block.header.parentHash.toString('hex')) {
-      await this.chain.putBlocks([block])
-    } else {
-      // If block is beyond current tip, handle as `NEW_BLOCK_HASHES`
-      this.handleNewBlockHashes([[block.header.hash(), block.header.number]])
-    }
-  }
-
   async stop(): Promise<boolean> {
     if (!this.running) {
       return false
